+++ workspaces/jenkins/packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [adfc3bb]
+  - @backstage-community/plugin-jenkins@0.10.2
+
 ## 0.0.2
 
 ### Patch Changes