+++ workspaces/scaffolder-relation-processor/plugins/catalog-backend-module-scaffolder-relation-processor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage-community/plugin-catalog-backend-module-scaffolder-relation-processor
 
+## 1.2.3
+
+### Patch Changes
+
+- 7b55dc3: Backstage version bump to v1.29.2
+
+## 1.2.2
+
+### Patch Changes
+
+- b3615da: update README.md to reference the correct package name
+
+## 1.2.1
+
+### Patch Changes
+
+- 8288c9c: update package to backstage 1.27.7 and bump version to a version number sematically larger than 1.1.x to replace the janus-idp package.
+
 ## 0.2.0
 
 ### Minor Changes