+++ workspaces/tech-insights/packages/backend/CHANGELOG.md
@@ -1,5 +1,25 @@
 # backend
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [e49b4eb]
+  - @backstage-community/plugin-tech-insights-backend@0.6.0
+  - app@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [7ac338c]
+- Updated dependencies [a8d8d44]
+- Updated dependencies [794cc8b]
+  - @backstage-community/plugin-tech-insights-backend-module-jsonfc@0.1.53
+  - @backstage-community/plugin-tech-insights-backend@0.5.35
+  - @backstage-community/plugin-tech-insights-node@0.6.4
+  - app@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes